requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.0
idna==3.6
multidict==6.0.4
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
urllib3==2.1.0
yarl==1.9.3
